+++ controllers/package.go
@@ -1,16 +1,24 @@
 package controllers
 
 import (
+	"context"
 	"strings"
 
 	gsannotations "github.com/giantswarm/k8smetadata/pkg/annotation"
+	"github.com/pkg/errors"
+	"k8s.io/apimachinery/pkg/types"
 	capa "sigs.k8s.io/cluster-api-provider-aws/api/v1beta1"
 	eks "sigs.k8s.io/cluster-api-provider-aws/controlplane/eks/api/v1beta1"
+	capi "sigs.k8s.io/cluster-api/api/v1beta1"
 
 	"github.com/aws-resolver-rules-operator/pkg/resolver"
 )
 
 //go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 -generate
+
+const (
+	DnsFinalizer = "capa-operator.finalizers.giantswarm.io/dns-controller"
+)
 
 func buildClusterFromAWSCluster(awsCluster *capa.AWSCluster, identity *capa.AWSClusterRoleIdentity, mcIdentity *capa.AWSClusterRoleIdentity) resolver.Cluster {
 	cluster := resolver.Cluster{
@@ -73,4 +81,41 @@ func getSubnetIds(subnets capa.Subnets) []string {
 	}
 
 	return subnetIds
+}
+
+// getBastionIp tries to find a bastion machine in this cluster and fetch its IP address from the status field.
+// It will return the internal IP address when using private VPC mode, or an external IP address otherwise.
+func getBastionIp(ctx context.Context, clusterClient ClusterClient, cluster resolver.Cluster) (string, error) {
+	bastionMachine, err := clusterClient.GetBastionMachine(ctx, cluster.Name)
+	if err != nil {
+		return "", errors.WithStack(err)
+	}
+
+	addressType := capi.MachineExternalIP
+	if cluster.IsVpcModePrivate {
+		addressType = capi.MachineInternalIP
+	}
+
+	for _, addr := range bastionMachine.Status.Addresses {
+		if addr.Type == addressType {
+			return addr.Address, nil
+		}
+	}
+
+	return "", nil
+}
+
+//counterfeiter:generate . ClusterClient
+type ClusterClient interface {
+	GetAWSCluster(context.Context, types.NamespacedName) (*capa.AWSCluster, error)
+	GetAWSManagedControlPlane(context.Context, types.NamespacedName) (*eks.AWSManagedControlPlane, error)
+	GetCluster(context.Context, types.NamespacedName) (*capi.Cluster, error)
+	AddAWSClusterFinalizer(ctx context.Context, cluster *capa.AWSCluster, finalizer string) error
+	AddAWSManagedControlPlaneFinalizer(ctx context.Context, awsManagedControlPlane *eks.AWSManagedControlPlane, finalizer string) error
+	Unpause(context.Context, *capa.AWSCluster, *capi.Cluster) error
+	RemoveAWSClusterFinalizer(ctx context.Context, awsCluster *capa.AWSCluster, finalizer string) error
+	RemoveAWSManagedControlPlaneFinalizer(ctx context.Context, awsManagedControlPlane *eks.AWSManagedControlPlane, finalizer string) error
+	GetIdentity(context.Context, *capa.AWSIdentityReference) (*capa.AWSClusterRoleIdentity, error)
+	MarkConditionTrue(context.Context, *capi.Cluster, capi.ConditionType) error
+	GetBastionMachine(ctx context.Context, clusterName string) (*capi.Machine, error)
 }