+++ main.go
@@ -142,6 +142,11 @@ func main() {
 		os.Exit(1)
 	}
 
+	if err = (controllers.NewEKSDnsReconciler(k8sClusterClient, dns, managementClusterName, managementClusterNamespace)).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create controller")
+		os.Exit(1)
+	}
+
 	// +kubebuilder:scaffold:builder
 
 	if err := mgr.AddHealthzCheck("healthz", healthz.Ping); err != nil {