+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Reconcile `AWSClusters` and `AWSManagedControlPlane` instead of `Cluster` CR when reconciling dns records.
+
 ## [0.10.1] - 2023-10-03
 
 ### Changed